Fix warnings

+++ Mimetic/mimetic/rfc822/field.cxx
@@ -39,7 +39,7 @@ Field::Field(const string& line)
     {
         m_name.assign(line.begin(), line.begin() + colon);
         unsigned int i;
-        for(i = 1 + colon; i < line.length() - 1 && line[i] == ' '; ++i)
+        for(i = 1 + (unsigned int)colon; i < line.length() - 1 && line[i] == ' '; ++i)
             ; // skip spaces before field-body
         string val(line.begin() +i, line.end());
         value(val);
@@ -158,7 +158,7 @@ ostream& Field::write(ostream& os, unsigned int fold) const
         string ostr = name() + ": " + value();
 
         // skip the "fieldname: " part just on the first inner iteration 
-        skip = name().length() + 2; 
+        skip = (int)name().length() + 2;
 
         while(ostr.length() > fold)
         {
